add debuger
