+++ python/paddle/v2/fluid/framework.py
@@ -1122,59 +1122,3 @@ def program_guard(main_program, startup_program=None):
     switch_main_program(main_program)
     if startup_program is not None:
         switch_startup_program(startup_program)
-
-
-def debug_block_image(block, highlights=[], path="./temp.dot"):
-    '''
-    Generate a debug graph for block.
-    Args:
-        block(Block): a block.
-    '''
-    from graphviz import GraphPreviewGenerator
-    graph = GraphPreviewGenerator("some graph")
-    # collect parameters and args
-    protostr = block.desc.serialize_to_string()
-    desc = framework_pb2.BlockDesc.FromString(str(protostr))
-
-    def need_highlight(name):
-        for pattern in highlights:
-            assert type(pattern) is str
-            if re.match(pattern, name):
-                return True
-        return False
-
-    # draw parameters and args
-    vars = {}
-    for var in desc.vars:
-        shape = [str(i) for i in var.lod_tensor.tensor.dims]
-        if not shape:
-            shape = ['null']
-        # create var
-        if var.persistable:
-            varn = graph.add_param(
-                var.name, var.type, shape, highlight=need_highlight(var.name))
-        else:
-            varn = graph.add_arg(var.name, highlight=need_highlight(var.name))
-        vars[var.name] = varn
-
-    def add_op_link_var(op, var, op2var=False):
-        for arg in var.arguments:
-            if arg not in vars:
-                # add missing variables as argument
-                vars[arg] = graph.add_arg(arg, highlight=need_highlight(arg))
-            varn = vars[arg]
-            highlight = need_highlight(op.description) or need_highlight(
-                varn.description)
-            if op2var:
-                graph.add_edge(op, varn, highlight=highlight)
-            else:
-                graph.add_edge(varn, op, highlight=highlight)
-
-    for op in desc.ops:
-        opn = graph.add_op(op.type, highlight=need_highlight(op.type))
-        for var in op.inputs:
-            add_op_link_var(opn, var, False)
-        for var in op.outputs:
-            add_op_link_var(opn, var, True)
-
-    graph(path, show=True)
